refactor(settings): remove redundant setup surfaces
